pyproject.toml
[tool.poetry]
name = "applicationName"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
win11toast = "^0.35"
bottle = "^0.13.2"
infi-systray = "^0.1.12.1"

[tool.poetry.dev-dependencies]


[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.11.1"
imageio = "^2.36.1"
nuitka = "^2.5.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
